Option to handle right and middle-click
I was making some entities earlier and I wanted a right-click handler. I ended up making my own little `onClick` variable and doing fancy stuff with the input handler for my entities which could receive clicks...

I feel like right and middle-click should be handled by the engine gracefully without needing to do extra stuff.
In this example, the on_click attribute can now be a tuple, and if it is a tuple, right and middle clicks can be used.

Right click checks to see if `on_click` is a tuple and if it is, makes sure it's callable (not None. This is so that a middle-click handler can be set without a right-click handler existing).
The middle click checks to see if `on_click` is a tuple and if it contains enough elements to have an entry for middle-click (len > 2)